Use the same queries for pruning id filter tables and tx meta table
